Move rest of test functions
